requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
boto3==1.34.54
    # via django-ses
botocore==1.34.54
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
crispy-bootstrap5==2024.2
    # via -r requirements.in
cryptography==42.0.5
    # via social-auth-core
dawg2==0.13.0
    # via -r requirements.in
decorator==5.1.1
    # via ipython
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.10
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-appconf
    #   django-crispy-forms
    #   django-extensions
    #   django-ses
    #   django-storages
    #   djangorestframework
    #   social-auth-app-django
django-appconf==1.0.6
    # via django-compressor
django-compressor==4.4
    # via -r requirements.in
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-react-components==0.1.8a0
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
django-sass-processor==1.4
    # via -r requirements.in
django-ses==3.5.2
    # via -r requirements.in
django-storages==1.14.2
    # via -r requirements.in
django-webpack-loader==0.7.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via django-rest-framework
ec2-metadata==2.13.0
    # via -r requirements.in
executing==2.0.1
    # via stack-data
idna==3.6
    # via requests
ipython==8.22.1
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
libsass==0.23.0
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via ipython
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.43
    # via ipython
psycopg2==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via ipython
pyjwt==2.8.0
    # via social-auth-core
python-dateutil==2.9.0.post0
    # via botocore
python3-openid==3.2.0
    # via social-auth-core
pytz==2024.1
    # via
    #   django-ses
    #   djangorestframework
rcssmin==1.1.1
    # via django-compressor
requests==2.31.0
    # via
    #   -r requirements.in
    #   ec2-metadata
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rjsmin==1.2.1
    # via django-compressor
s3transfer==0.10.0
    # via boto3
sentry-sdk==1.40.6
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
social-auth-app-django==5.4.0
    # via -r requirements.in
social-auth-core==4.5.3
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
urllib3==2.0.7
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wcwidth==0.2.13
    # via prompt-toolkit
